Delivery Module/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
supabase==2.18.1
httpx==0.28.1
boto3==1.39.9
pydantic-settings==2.10.1
